requirements.txt
bert-extractive-summarizer
bs4
openai
pymongo
dnspython
requests
spacy
torch
tornado
transformers
